requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
httpx
slowapi
fastapi-cache2
jinja2

# Optional, but recommended for multi-worker deployments
# to share cache and rate-limit state effectively:
# redis>=4.0.0